+++ src/relax/op/nn/nn.h
@@ -49,7 +49,7 @@ namespace relax {
 Expr relu(Expr data);
 
 /*! \brief Leaky rectified linear unit. */
-Expr leakyrelu(Expr data, float alpha);
+Expr leakyrelu(Expr data, double alpha);
 
 /*! \brief Gaussian Error Linear Units function. */
 Expr gelu(Expr data);
